// src/app/enums/fiat-currencies.ts
export enum FiatCurrencies {
  usd = 'USD',
  eur = 'EUR',
  brl = 'BRL',
  rub = 'RUB',
  uah = 'UAH',
  ngn = 'NGN',
  cny = 'CNY',
  vnd = 'VND',
  inr = 'INR',
  gbp = 'GBP',
  jpy = 'JPY',
  cad = 'CAD',
  php = 'PHP',
  zar = 'ZAR',
  pkr = 'PKR',
  thb = 'THB',
  try = 'TRY',
  twd = 'TWD',
  mxn = 'MXN',
  mad = 'MAD',
  idr = 'IDR',
  bdt = 'BDT',
  ars = 'ARS',
}