// src/app/types/currenciesApiResponse.ts
export type CurrenciesApiResponse = {
    date: string
    usd: Usd
  }